+++ sdk/search/search/src/index.ts
@@ -57,6 +57,7 @@ export {
 export { SearchApiKeyCredential } from "./searchApiKeyCredential";
 export { default as GeographyPoint } from "./geographyPoint";
 export { odata } from "./odata";
+export { IndexDocumentsBatch } from "./indexDocumentsBatch";
 export {
   AutocompleteResult,
   AutocompleteMode,