+++ sdk/search/search/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 11.0.0-preview.2 (Unreleased)
 
+- Support for index management operations via the `SearchServiceClient`.
+- [Breaking] `indexDocuments` on `SearchIndexClient` now takes an `IndexDocumentsBatch` object instead of a raw action array. This new type helps compose an array of actions to be performed on the index.
 
 ## 11.0.0-preview.1 (2020-03-09)